mv content roles consts, get reviewer to utils
