+++ .gitlab/ci/.gitlab-ci.sdk-nightly.yml
@@ -231,7 +231,7 @@ demisto-sdk-nightly:run-end-to-end-tests:
   script:
     - |
       mkdir $ARTIFACTS_FOLDER/demisto-sdk
-      git clone -b master --single-branch --depth 1 https://github.com/demisto/demisto-sdk.git $ARTIFACTS_FOLDER/demisto-sdk
+      git clone -b master --single-branch --depth 1 https://github.com/demisto/demisto-sdk.git@${SDK_REF} $ARTIFACTS_FOLDER/demisto-sdk
 
       TUNNEL_PORT=$(cat $ARTIFACTS_FOLDER/env_results.json | jq "map(select(.Role == \"$INSTANCE_ROLE\")) | .[].TunnelPort")
       export DEMISTO_BASE_URL="https://$(cat $ARTIFACTS_FOLDER/env_results.json | jq -r '.[0].InstanceDNS')"
